add taskqueue for pioneer which is more stable with multiple request, e.g. many dockers at the same time fix error of buffer overflow when invoke system commands add version and build time for salsa metadata increase memory for pioneer at bootstrap use thread for message handling which block the pioneer

+++ salsa-core-pom/salsa-engine/src/main/java/at/ac/tuwien/dsg/cloud/salsa/engine/impl/richInformationCapability/RichInformationUnitCapability.java
@@ -119,8 +119,12 @@ public class RichInformationUnitCapability implements UnitCapabilityInterface {
         ServiceUnit unit = service.getComponentById(nodeId);
         ServiceInstance instance = unit.getInstanceById(instanceId);
 
-        unitInstanceDAO.deleteUnitInstanceByID(instance.getUuid().toString());
-        lowerCapa.remove(serviceId, nodeId, instanceId);
+        if (instance != null) {
+            if (unitInstanceDAO != null) {
+                unitInstanceDAO.deleteUnitInstanceByID(instance.getUuid().toString());
+            }
+            lowerCapa.remove(serviceId, nodeId, instanceId);
+        }
     }
 
     private UnitInstance makeUnitInstance(CloudService service, ServiceUnit unit, ServiceInstance ins) {
